류진아 : Bullet 기능 및 ImageCut 기능 업데이트

+++ API/GameEngineBase/GameEngineTime.cpp
@@ -25,7 +25,7 @@ void GameEngineTime::Update()
 {
 	QueryPerformanceCounter(reinterpret_cast<LARGE_INTEGER*>(&CurrentCount_));
 
-	RealDeltaTime_ = static_cast<double>((CurrentCount_ - PrevCount_) / SecondCount_);
+	RealDeltaTime_ = (static_cast<double>(CurrentCount_) - static_cast<double>(PrevCount_)) / static_cast<double>(SecondCount_);
 	PrevCount_ = CurrentCount_;
 
 	DeltaTime_ = static_cast<float>(RealDeltaTime_);
